--- encodec_16k_lanch/modules/__init__.py
@@ -1,22 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-
-"""Torch modules."""
-
-# flake8: noqa
-from .conv import (
-    pad1d,
-    unpad1d,
-    NormConv1d,
-    NormConvTranspose1d,
-    NormConv2d,
-    NormConvTranspose2d,
-    SConv1d,
-    SConvTranspose1d,
-)
-from .lstm import SLSTM
-from .seanet import SEANetEncoder, SEANetDecoder
-from .transformer import StreamingTransformerEncoder
