--- encodec_16k_lanch/test.sh
@@ -1,4 +1,0 @@
-python test.py "/test" \
-  "./output_ep15" \
-  --resume_path ./model_path/2023-03-25-18-15/best_15.pth
-
